requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
dj-database-url
django-environ
djangorestframework-simplejwt
django-cors-headers
drf_yasg
